+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dependency injector
 - `tempo:attributes` command to save extra attributes for logging
 - `tempo:list` command to list saved tempo attributes
+- Save tempo group on log start
+- Save tempo group on log stop
 
 ## [0.2.2] - 2020-03-11
 ### Fixed